Rename function.
